Add @types/styled-system, Move Header from /containers to /components, Update bio
