+++ v5/package.json
@@ -26,6 +26,7 @@
     "@types/node": "^13.13.5",
     "@types/react": "^16.9.34",
     "@types/styled-components": "^5.1.3",
+    "@types/styled-system": "^5.1.10",
     "babel-plugin-styled-components": "^1.8.0",
     "typescript": "^4.0.3"
   },
